Improve cluster script & HCP pipeline

+++ scripts/cluster-pipeline-executor.sh
@@ -12,7 +12,7 @@ CPUS_PER_TASK=15
 ################################################################
 
 
-while getopts 'i:o:t:l:p:r:m:n:b:d:c:h' opt; do
+while getopts 'i:o:t:l:p:r:m:n:b:d:c:s:h' opt; do
   case "$opt" in
     i) INDIR="$OPTARG";;
     o) OUTDIR="$OPTARG";;
@@ -25,11 +25,12 @@ while getopts 'i:o:t:l:p:r:m:n:b:d:c:h' opt; do
     b) BRANCH="$OPTARG";;
     d) SUBMIT_DELAY="$OPTARG";;
     c) CPUS_PER_TASK="$OPTARG";;
+    s) SIF_PATH="$OPTARG";;
 
     ?|h)
       echo "-i      Input BIDS dataset"
       echo "-o      Derivatives dir (i.e., where to store the results)"
-      echo "-t      Where to store temporary PUMI workflow files on the cluster (MUST BE SOMEWHERE IN /tmp !)"
+      echo "-t      Where to store temporary PUMI workflow files on the worker nodes (MUST BE an absolute path somewhere in /local/)"
       echo "-l      NFS directory that should be used to store the Slurm log files (+ Apptainer SIF file)"
       echo "-p      PUMI pipeline you want to run (default: '${PIPELINE}')"
       echo "-r      Nipype plugin params to limit resource usage (default: '${RESOURCES}')"
@@ -38,6 +39,7 @@ while getopts 'i:o:t:l:p:r:m:n:b:d:c:h' opt; do
       echo "-b      Which PUMI GitHub branch to install (default: '${BRANCH}')"
       echo "-d      Minimum delay between submission of jobs in seconds (default: '${SUBMIT_DELAY}')"
       echo "-c      CPU's per task (default: '${CPUS_PER_TASK}')"
+      echo "-s      Path to Image SIF file. If not provided, pull and convert latest docker image."
       exit 1
       ;;
   esac
@@ -55,6 +57,7 @@ echo "Max Slurm jobs of user (-m): ${MAX_JOBS}"
 echo "Slurm nice value (-n): ${NICE}"
 echo "PUMI branch to use (-b): ${BRANCH}"
 echo "Minimum delay between submission of jobs in seconds (-d): ${SUBMIT_DELAY}"
+echo "Path to Image SIF file (-s): ${SIF_PATH}"
 echo "--------------------------------------------------------------------"
 
 # Validation to ensure mandatory options are provided.
@@ -65,6 +68,15 @@ if [ -z "${INDIR}" ] || [ -z "${OUTDIR}" ] || [ -z "${TMP_PUMI}" ] || [ -z "${LO
   exit 1
 fi
 
+if [[ "${TMP_PUMI}" == /local/* ]];
+then
+  echo "-t was specified to a sub-directory in /local. Great!"
+else
+  echo "Error: -t must be set to a sub-directory in /local/"
+  exit 1
+
+fi
+
 ############################# Main script begins here #########################################
 
 dataset_name=$(basename "$INDIR")
@@ -72,13 +84,17 @@ mkdir -p "jobs_scripts/${dataset_name}"  # Create directory which will contain t
 
 mkdir -p "${LOG_PATH}"  # Create directory where the jobs will store the slurm outputs in
 
-# Pull (and convert) PUMI image locally and then copy to NFS share where all the jobs can copy it from
-rm -rf ${TMP_PUMI}/apptainer_cache/
-mkdir -p ${TMP_PUMI}/apptainer_cache/
-APPTAINER_CACHEDIR=${TMP_PUMI}/apptainer_cache/ \
- apptainer pull ${TMP_PUMI}/PUMI.sif docker://pnilab/pumi:latest
-cp ${TMP_PUMI}/PUMI.sif ${LOG_PATH}/PUMI.sif
-rm -rf ${TMP_PUMI}
+if [ -z "${SIF_PATH}" ]; then
+  rm -rf ${TMP_PUMI}/apptainer_cache/
+  mkdir -p ${TMP_PUMI}/apptainer_cache/
+  # Pull (and convert) PUMI image locally and then copy to NFS share where all the jobs can copy it from
+  APPTAINER_CACHEDIR=${TMP_PUMI}/apptainer_cache/ apptainer pull ${TMP_PUMI}/PUMI.sif docker://pnilab/pumi:latest
+  cp ${TMP_PUMI}/PUMI.sif ${LOG_PATH}/PUMI.sif
+  rm -rf ${TMP_PUMI}
+  SIF_PATH=${LOG_PATH}/PUMI.sif
+else
+  echo "SIF was already provided. No pulling and conversion needed."
+fi
 
 mkdir -p "${OUTDIR}"
 
@@ -102,36 +118,40 @@ echo "*************************************************************"
 echo "Starting on \$(hostname) at \$(date +"%T")"
 echo "*************************************************************"
 
-subject_data_in="${TMP_PUMI}/input/${subject_id}" # Create temporary directory which stores BIDS data for one subject
+subject_dir="${TMP_PUMI}/${subject_id}/"
+
+subject_data_in="\${subject_dir}/input/" # Create temporary directory which stores BIDS data for one subject
 rm -rf "\${subject_data_in}"
 mkdir -p "\${subject_data_in}"
 
-subject_data_out="${TMP_PUMI}/output/${subject_id}" # Create temporary directory which stores derivatives for one subject
+subject_data_out="\${subject_dir}/output/" # Create temporary directory which stores derivatives for one subject
 rm -rf "\${subject_data_out}"
 mkdir -p "\${subject_data_out}"
 
-subject_tmp="${TMP_PUMI}/tmp/${subject_id}" # Create temporary directory which stores derivatives for one subject
+subject_tmp="\${subject_dir}/tmp/" # Create temporary directory which stores derivatives for one subject
 rm -rf "\${subject_tmp}"
 mkdir -p "\${subject_tmp}"
 
 cp -vr "${subject_folder}" "\${subject_data_in}"
 cp -v "${dataset_description_path}" "\${subject_data_in}"  # Every valid BIDS dataset must contain description (otherwise Nipype raises BIDSValidationError)
 
-pumi_dir=${TMP_PUMI}/PUMI/${subject_id}/
+pumi_dir="\${subject_dir}/PUMI/"
 rm -rf \${pumi_dir}
 mkdir -p \${pumi_dir}  # Create folder in which we clone PUMI into (and parent folders if necessary)
 
-mkdir -p ${TMP_PUMI}/apptainer_image/${subject_id}/
-cp ${LOG_PATH}/PUMI.sif ${TMP_PUMI}/apptainer_image/${subject_id}/PUMI.sif
+apptainer_image_dir="\${subject_dir}/apptainer_image/"
+apptainer_image="\${apptainer_image_dir}/PUMI.sif"
+mkdir -p "\${apptainer_image_dir}"
+cp ${SIF_PATH} "\${apptainer_image}"
 
-subject_apptainer_cache_dir=${TMP_PUMI}/apptainer_cache/${subject_id}/
+subject_apptainer_cache_dir="\${subject_dir}/apptainer_app_cache/"
 rm -rf \${subject_apptainer_cache_dir}
 mkdir -p \${subject_apptainer_cache_dir}
 
 APPTAINER_CACHEDIR=\${subject_apptainer_cache_dir} \
 apptainer exec \
 --writable-tmpfs \
-${TMP_PUMI}/apptainer_image/${subject_id}/PUMI.sif \
+\${apptainer_image} \
 bash -c " \
 set -x; \
 git clone -b ${BRANCH} https://github.com/pni-lab/PUMI \${pumi_dir}; \
@@ -156,10 +176,7 @@ cp -vr \${subject_data_out}/* ${OUTDIR}/
 
 
 # Remove (most) files from cluster
-rm -rf "\${subject_data_in}"
-rm -rf "\${subject_data_out}"
-rm -rf "\${subject_tmp}"
-rm -rf "\${subject_apptainer_cache_dir}"
+rm -rf "\${subject_dir}"
 
 echo "*************************************************************"
 echo "Ended on \$(hostname) at \$(date +"%T")"
